Add alias module This removes alias code from tell into its own module
